Rework playbooks to be used locally
Since I did not use this set of scripts for a long time, it is not
working on a local machine (in a few words, Ansible changed).
Additionally, I just realized that I did not take proper care related to
my changes and the playbooks got messed. This commit makes the playbooks
work locally again and clean up some stuff.

Signed-off-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>